+++ apstools/_devices/__init__.py
@@ -4,6 +4,25 @@ from .aps_cycle import ApsCycleDM
 from .aps_machine import ApsMachineParametersDevice
 from .aps_undulator import ApsUndulator
 from .aps_undulator import ApsUndulatorDual
+from .area_detector_support import AD_FrameType_schemes
+from .area_detector_support import AD_plugin_primed
+from .area_detector_support import AD_prime_plugin
+from .area_detector_support import AD_prime_plugin2
+from .area_detector_support import AD_EpicsHdf5FileName
+from .area_detector_support import AD_EpicsJpegFileName
+from .axis_tuner import AxisTunerException
+from .axis_tuner import AxisTunerMixin
+from .description_mixin import EpicsDescriptionMixin
+from .kohzu_monochromator import KohzuSeqCtl_Monochromator
+from .mixin_base import DeviceMixinBase
+from .motor_mixins import EpicsMotorDialMixin
+from .motor_mixins import EpicsMotorEnableMixin
+from .motor_mixins import EpicsMotorLimitsMixin
+from .motor_mixins import EpicsMotorRawMixin
+from .motor_mixins import EpicsMotorResolutionMixin
+from .motor_mixins import EpicsMotorServoMixin
+from .scaler_support import SCALER_AUTOCOUNT_MODE
+from .scaler_support import use_EPICS_scaler_channels
 from .shutters import ApsPssShutter
 from .shutters import ApsPssShutterWithStatus
 from .shutters import EpicsMotorShutter
@@ -11,4 +30,6 @@ from .shutters import EpicsOnOffShutter
 from .shutters import OneSignalShutter
 from .shutters import ShutterBase
 from .shutters import SimulatedApsPssShutterWithStatus
+from .struck3820 import Struck3820
 from .tracking_signal import TrackingSignal
+from .xia_pf4 import DualPf4FilterBox
