+++ apstools/_devices/mixin_base.py
@@ -1,0 +1,18 @@
+"""
+Base class for Device Mixins
++++++++++++++++++++++++++++++++++++++++
+
+.. autosummary::
+
+   ~DeviceMixinBase
+"""
+
+from ophyd import Device
+
+
+class DeviceMixinBase(Device):
+    """
+    Base class for apstools Device mixin classes
+
+    .. index:: Ophyd Device Mixin; DeviceMixinBase
+    """
