+++ apstools/_devices/xia_pf4.py
@@ -1,0 +1,44 @@
+"""
+XIA PF4 Filters
++++++++++++++++++++++++++++++++++++++++
+
+.. autosummary::
+
+   ~DualPf4FilterBox
+"""
+
+from ophyd import Component
+from ophyd import Device
+from ophyd import EpicsSignal
+from ophyd import EpicsSignalRO
+
+
+class DualPf4FilterBox(Device):
+    """
+    Dual Xia PF4 filter boxes using support from synApps (using Al, Ti foils)
+
+    .. index:: Ophyd Device; DualPf4FilterBox
+
+    EXAMPLE::
+
+        pf4 = DualPf4FilterBox("2bmb:pf4:", name="pf4")
+        pf4_AlTi = DualPf4FilterBox("9idcRIO:pf4:", name="pf4_AlTi")
+
+    """
+
+    fPosA = Component(EpicsSignal, "fPosA")
+    fPosB = Component(EpicsSignal, "fPosB")
+    bankA = Component(EpicsSignalRO, "bankA")
+    bankB = Component(EpicsSignalRO, "bankB")
+    bitFlagA = Component(EpicsSignalRO, "bitFlagA")
+    bitFlagB = Component(EpicsSignalRO, "bitFlagB")
+    transmission = Component(EpicsSignalRO, "trans")
+    transmission_a = Component(EpicsSignalRO, "transA")
+    transmission_b = Component(EpicsSignalRO, "transB")
+    inverse_transmission = Component(EpicsSignalRO, "invTrans")
+    thickness_Al_mm = Component(EpicsSignalRO, "filterAl")
+    thickness_Ti_mm = Component(EpicsSignalRO, "filterTi")
+    thickness_glass_mm = Component(EpicsSignalRO, "filterGlass")
+    energy_keV_local = Component(EpicsSignal, "E:local")
+    energy_keV_mono = Component(EpicsSignal, "displayEnergy")
+    mode = Component(EpicsSignal, "useMono", string=True)
