+++ docs/source/source/_devices.rst
@@ -15,6 +15,9 @@ Submodules
 .. automodule:: apstools._devices.aps_bss_user
     :members:
 
+.. automodule:: apstools._devices.area_detector_support
+    :members:
+
 .. automodule:: apstools._devices.aps_cycle
     :members:
 
@@ -24,8 +27,32 @@ Submodules
 .. automodule:: apstools._devices.aps_undulator
     :members:
 
+.. automodule:: apstools._devices.axis_tuner
+    :members:
+
+.. automodule:: apstools._devices.description_mixin
+    :members:
+
+.. automodule:: apstools._devices.kohzu_monochromator
+    :members:
+
+.. automodule:: apstools._devices.mixin_base
+    :members:
+
+.. automodule:: apstools._devices.motor_mixins
+    :members:
+
+.. automodule:: apstools._devices.scaler_support
+    :members:
+
 .. automodule:: apstools._devices.shutters
     :members:
 
+.. automodule:: apstools._devices.struck3820
+    :members:
+
 .. automodule:: apstools._devices.tracking_signal
+    :members:
+
+.. automodule:: apstools._devices.xia_pf4
     :members:
